feat(collapse): migrate to ts
