+++ packages/collapse/src/Body.spec.tsx
@@ -1,5 +1,5 @@
+import * as React from 'react';
 import { shallow } from 'enzyme';
-import React from 'react';
 import Body from './Body';
 
 describe('Accordion.Body', () => {
