+++ packages/collapse/src/Constant.ts
