+++ packages/collapse/src/index.ts
