+++ packages/collapse/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tsconfig.dev.json",
+  "include": ["./src/index.ts"],
+}
